refactor: move logic related to proxies into wallet proxies model, connect to ui directly via gate
