+++ src/renderer/features/wallet-details/ui/wallets/ProxiedWalletDetails.tsx
@@ -16,6 +16,7 @@ import { AddPureProxied } from '@/features/proxied-add-pure';
 import { AddProxy } from '@/features/proxy-add';
 import { RenameWallet } from '@/features/wallets/RenameWallet';
 import { walletDetailsModel } from '../../model/wallet-details-model';
+import { walletProxiesModel } from '../../model/wallet-proxies-model';
 import { WalletFiatBalance } from '../components';
 import { ProxiesList } from '../components/ProxiesList';
 import { Action, type WalletAction, WalletActions } from '../components/WalletActions';
@@ -41,13 +42,14 @@ type Props = {
 
 export const ProxiedWalletDetails = ({ wallet, onClose }: Props) => {
   useGate(walletDetailsModel.flow, { wallet });
+  useGate(walletProxiesModel.flow, { wallet });
 
   const { t } = useI18n();
 
   const chains = useUnit(networkModel.$chains);
   const wallets = useUnit(walletModel.$wallets);
-  const hasProxies = useUnit(walletDetailsModel.$hasProxies);
-  const proxiesCount = useUnit(walletDetailsModel.$proxiesCount);
+  const hasProxies = useUnit(walletProxiesModel.$hasWalletProxies);
+  const proxiesCount = useUnit(walletProxiesModel.$walletProxiesCount);
   const canCreateProxy = useUnit(walletDetailsModel.$canCreateProxy);
 
   const [isRenameInputOpen, toggleIsRenameInputOpen] = useToggle();
