+++ src/renderer/features/wallet-details/ui/wallets/WatchOnlyWalletDetails.tsx
@@ -15,6 +15,7 @@ import { accountUtils, walletUtils } from '@/entities/wallet';
 import { ForgetWalletConfirm } from '@/features/wallets/ForgetWallet';
 import { RenameWallet } from '@/features/wallets/RenameWallet';
 import { walletDetailsModel } from '../../model/wallet-details-model';
+import { walletProxiesModel } from '../../model/wallet-proxies-model';
 import { WalletFiatBalance } from '../components';
 import { ProxiesList } from '../components/ProxiesList';
 
@@ -26,12 +27,13 @@ type Props = {
 };
 export const WatchOnlyWalletDetails = ({ wallet, onClose }: Props) => {
   useGate(walletDetailsModel.flow, { wallet });
+  useGate(walletProxiesModel.flow, { wallet });
   const { t } = useI18n();
 
   const allChains = useUnit(networkModel.$chains);
-  const hasProxies = useUnit(walletDetailsModel.$hasProxies);
+  const hasProxies = useUnit(walletProxiesModel.$hasWalletProxies);
   const canCreateProxy = useUnit(walletDetailsModel.$canCreateProxy);
-  const proxiesCount = useUnit(walletDetailsModel.$proxiesCount);
+  const proxiesCount = useUnit(walletProxiesModel.$walletProxiesCount);
 
   const firstAccount = useStoreMap({
     store: accounts.$list,
